LayoutTests/fast/mediastream/applyConstraints-ended.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <script src="../../resources/gc.js"></script>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
</head>
<body>
    <script>
    promise_test(async (t) => {
        const stream = await navigator.mediaDevices.getUserMedia({ video: true });
        const track = stream.getVideoTracks()[0];

        const badConstraints = { width: { exact: 10000000  } };
        await track.applyConstraints(badConstraints).then(assert_unreached, e => {
            assert_equals(e.name, "OverconstrainedError");
        });
        track.stop();
        await track.applyConstraints(badConstraints);
    }, "applyConstraints on an ended video track");

    promise_test(async (t) => {
        const stream = await navigator.mediaDevices.getUserMedia({ audio: true });
        const track = stream.getAudioTracks()[0];

        const badConstraints = { sampleRate: { exact: 10000000  } };
        await track.applyConstraints(badConstraints).then(assert_unreached, e => {
            assert_equals(e.name, "OverconstrainedError");
        });
        track.stop();
        await track.applyConstraints(badConstraints);
    }, "applyConstraints on an ended audio track");
    </script>
</body>
</html>
